refactor: Edit sql lite database module
