+++ core/data/build.gradle.kts
@@ -10,4 +10,5 @@ dependencies {
     implementation(projects.core.domain)
     implementation(projects.core.model)
     implementation(projects.core.datastore)
+    implementation(projects.core.database)
 }
