+++ core/database/build.gradle.kts
@@ -10,5 +10,4 @@ android {
 
 dependencies {
     implementation(projects.core.model)
-    implementation(projects.core.data)
 }
