+++ core/datastore/src/main/java/org/sopt/at/datastore/datastore/DefaultAutoSignInPreferencesDataSource.kt
@@ -6,6 +6,7 @@ import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
+import org.sopt.at.datastore.source.AutoSignInPreferencesDataSource
 import javax.inject.Inject
 import javax.inject.Named
